Add CC0 license.
Based on investigations by Sven Geggus, the majority of contributors to the
original OSM stylesheets are in favour of a CC0 license, or have not
contributed towards the cartography brought forward into this project.

If anyone can identify their copyrightable contributions within this project
that they do not wish licensed under CC0, then I'll happily and immediately
remove them.