Changes layer order
The layer order used to be as follows.

(bridges)                               B

I classified all layers between and including the bridge-layers and the tunnels-layer with one of the following classifications:

B    Bridge level.
R/B  Above roads but below bridges.
R    Road level.
T/R  Above tunnels but below roads.
T    Tunnel level.

I grouped all of the layers within one class together, without changing the order within the classes. For now, I ignored the buildings-layers, due to the difficulties with ordering buildings, highways and highway-areas described in #172.

This change causes a couple of changes in the rendering, which mainly occur in rare cases.
- Casings of turning circles are now drawn above foot/bike/track tunnels (they were already drawn above other tunnels), and above cliffs and barriers (just like other roads are drawn above cliffs and barriers).
- Landuse-overlays (i.e., military and national park), castlewalls, and citywalls are now drawn above foot/bike/track tunnels (they were already drawn above other tunnels).
- Ferry-routes are now drawn below roads (highways, tracks, turning-circles, and highway-areas).
- Railways on level 12 (due to the #roads-low-zoom layer), aerialways, and waterway-bridges are now drawn above access and directions markers of roads (they were already drawn above the roads themselves).