Render access restrictions at the same time as the roads to which they apply
Access restrictions are now rendered at the same time as the roads they apply to.
This prevents access restrictions from overlapping with roads they don't apply to.
This is accomplished by putting roads-fill and the bridge layers in the access class, and removing all separate access layers.

This affects rendering as follows:
Access restrictions are only displayed on top of the road they apply to, not on top of neighbouring roads.
This change does not affect rendering in any other way.