Add Arundina Sans to font list
Arundina Sans is based on DejaVu and has bold and italic versions, so it's an easy choice. Obtained with the ``fonts-sipa-arundina`` package.

This change makes mapnik/mapnik#2161 more obvious, but we were probably encountering that bug before with Unifont, it's just hard to tell with how hard Unifont was to read.