Render oneway roads narrower
Render oneway roads narrower than twoway roads for the types motorway, trunk, primary, secondary and tertiary.
This is accomplished by making twoway motorways wider and oneway trunk/primary/secondary/tertiary roads narrower.
More important road types are now always rendered wider than less important road types.

Changes:
- Twoway motorways are rendered wider
- Oneway trunk/primary/secondary/tertiary roads are rendered narrower
- Secondary is now narrower on z13/z14
- Tertiary_link is now rendered just like tertiary
- Round line caps on tertiary and larger are removed on regular roads to prevent #83
All other road types are rendered as before.

Further changes to the width of link roads are left for a future pull request.

I tried to achieve the goal of making more important road types wider than less important ones while at the same time making as little changes to the roadsizes as possible. I think it would be worth for if someone with feeling for design has a closer look at the sizes as we have them now. It should be very easy to adapt now: one can just change the variables in the beginning of the style.

I put the road sizes in a kind of table to make them more legible. If people think that's not practical, I can change it back to a list again.

This pull request resolves #265 and #83 on Github and 1952 on trac.