Remove special layering treatment sport areas
Sport areas are now ordered just like other landcover areas, i.e. by
z_order and way_area, instead of on top of other landuse.
In addition, pitches and tracks are no longer automatically rendered
on top of stadiums, as way_area should take care of that.

This resolves #394 on Github and #3224 on trac.