Add boat_sharing icon
Adding a icon for boat_sharing. Very much the same as car_sharing. (see
https://wiki.openstreetmap.org/wiki/Tag:amenity%3Dboat_sharing)

Rendering this does not interfer with other map POIs, since
boat_sharing is normally present in Marinas.
It is interesting for the general public to see it can share/rent a
boat.
Thanks for accepting this POI on the mapstyle.
Nounours