Order POI by way_area
POI are now ordered by way area. Note that icons are never rendered on top
of each other, so this has only influence on priority decision of icon
rendering.

Large POI (such as universities, malls, etc.) are now rendered with higher
priority.

The old order was:
* Render points with higher priority than areas
* Icon priority from high to low: tourism, amenity, highway, waterway,
  man_made, historic, shop, leisure
* Text priority from high to low: place, amenity, natural, historic, leisure,
  landuse, tourism, waterway, man_made, shop, aeroway
It is clear that the old order was quite a mess.

* Resolves partially #287
* Resolves partially #797
* Resolves #802
* Resolves https://trac.openstreetmap.org/ticket/4100