Reduce number of overlapping admin borders
This reduces the number of overlapping admin borders by rendering them in order
from high to low admin_level, and by giving all admin borders a white background
while using comp-op: darken.

Admin borders are currently grouped into three groups: level 1-4, level 5-8,
and level 9+. This changes only prevents overlap of admin borders within a
group (including overlap of borders of the same level). To prevent overlap of
borders within groups, it would be necessary to merge the groups, which would
lead to performance loss, and is therefore not included in this PR.

This partially solves #344.