Use a random symbology for forests
This adds a random symbology for forests, lightens forests, and
unifies natural=wood and landuse=forest rendering.

Using http://www.imagico.de/map/jsdotpattern.php a random pattern
was generated from the SVG

    <path d="m 3.5,0 -3.5,3.5 0,0.5 0.5,0 2.5,-2.5 0,2
    -3,3 0,0.5 0.5,0 2.5,-2.5 0,5.5
    1,0 0,-5.5 2.5,2.5 0.5,0 0,-0.5 -3,-3 0,-2
    2.5,2.5 0.5,0 0,-0.5 z" fill="rgb(58,135,39)"/>

    <path d="m 10.5,0 a 2.5,3 0 0,1 0,6 l 0,-1
    a 1.5,2 0 0,0 0,-4
    a 1.5,2 0 0,0 0,4 l 0,1
    a 2.5,3 0 0,1 0,-6
    z" fill="rgb(58,135,39)"/>

This was then converted with GIMP into a PNG which Mapnik can use.

The forest fill was lightened, but there are now darker symbols on.

This necessitated adjusting the text colour for forests, which was
done in Lch colour space. The frequent symbols also required some
halo adjustments.

There are multiple interpretations in use of landuse=forest vs.
natural=wood. Rather than attempt to sort this out when the tagging
has not settled, the same appearance is used for both.

This commit doesn't use different symbologies for different types
of forests (mixed/coniferous/broad_leaved/palm) (#822), but that
can be considered later.

Fixes #938
Fixes forest part of #937
Closes pnorman#6