Clean up low-zoom road definitions
This merges code for low-zoom road definitions into the corresponding
high-zoom road definitions.

Data on z<10 now comes from osm_planet_roads, data on z>=10 comes from
osm_planet_line. Chosen is for zoom level 10 as cut-off, because residential
is rendered from z10 and is not included in osm_planet_roads, so on z>=10
osm_planet_line needs to be queried anyway.

Advantages:
* This separates the zoomlevel cut-off for querying from the roads vs line
  table from the zoomlevel cut-off for the lowzoom rendering style.
* This brings related code (i.e. code per feature) closer together in the file,
  which should make the code more easy to modify.

This is mainly a code cleanup, but it causes some small changes in rendering:
* Take into account layering of roads on z10 and z11.
* Fix the loss of layering problem of railway tunnels on z12 (resolves #400).
* Fix bug that prevented tunnel rendering style for railway on z11/12, and add
  tunnel style rendering for spur/siding/yard for z<12.
* Give railways on z<12 a round line-join.

This is an improvement of #555.