Use a null check for shop in where clauses
Rather than checking a huge list that contains all common shop
values, we can just check for not null in the where clause. Because
the COALESCE checks the shop value any additional shop values
will come through as feature is null, and not match any rendering
rule. With the shop list encompassing all common values, this
will seldom happen.

This allows query simplification and should be no performance loss
and a potential gain.