render wayside_shrine the same way as wayside_cross
Wayside shrine is bigger and more visible in real life than
wayside_cross and it is shame that we display cross and not shrine.
Until somebody provides better icon, it is good enough to render it
as cross (as sometime it is called by local people cross anyway).

It is render only if religion is set to christian.
Other shrines need different icon.