Simplify WHERE clause of text layers
The WHERE clause is only needed for performance, as the filtering
is repeated by the COALESCE logic. Where we accept almost all rows
with a particular tag if that tag isn't null, we can use an
IS NOT NULL condition instead of a tag in ... construct, cutting
down on the length of the where clause.