Show national park boundaries from higher zoomlevel and larger size
With this commit, national park boundaries will be shown from the
same zoomlevel and size as their label, effectively eradicating
small unlabelled national parks from the rendering.

Such small national parks seem to contribute to a crowded feel of
the map in the areas and zoom levels where they occur.