Add tertiary to colour script; split up roads.mss
Tertiary roads are now the fifth and final road that has its colours
generated by generate_road_colours.py. The script is amended to
facilitate this.

The generated colour values now reside in road-colors-generated.mss, and
have been removed from roads.mms. road-colors-generated.mms is generated
by the colour script, and should not be edited manually. To regenerate
it:

```
./scripts/generate_road_colours.py > road-colors-generated.mss
```

The original output of the colour script included configuration values
not used by us directly, and provided some info about the amount of loss
caused by converting from CIE lch to RGB. To get this output as well as
the generated colours, call the script with the `-v` flag (for verbose).