Bring lightness of farmland color again in line with other landuse
Now that we have fading of landuse on z8-z12, I think we can revert
the fading of farmland (#1701). The reason for fading farmland was
mainly prominence of farmland on the mid-zoom levels, which is not
a problem anymore after the mid-zoom fading.

This brings the lightness of farmland again in line of the lightness
of similar landuses like grassland.