Select overground power=cable like power=line
Power cables with a location tag that says they are overground, are now
selected and rendered just like power lines, to improve the look of
substations that use gas insulated cables. Since those cables are
insulated, they have to be mapped with power=cable instead of
power=line, but this left undesirable gaps in the connections.

According to the wiki[1], cables without a location tag default to
'either underground or underwater', so they are treated as such.

[1] https://wiki.openstreetmap.org/wiki/Tag:power%3Dcable

Closes #3217.