Render administrative boundary text labels for relations only (#3762)
* Render text labels along borders of admin boundary relations only

Currently lines are only rendered for borders mapped as relations (i.e. osm_id <0) but the text labels along the borders are rendered for relations and closed ways. This commit adds a filter for "AND osm_id < 0" to the admin-text layer, so that the text labels do not appear in the rare occasions when an administrative boundary is mapped as a closed way instead of a relation

* Render central name labels for country and province/state relations only