Restore rendering of landcover text labels on points (#3878)
* Restore rendering of landcover and island text labels on points
* Select way_pixels as 0 if way_area is null

Prior to #3764 most features in `landcover` had text labels rendered at high zoom level when tagged on nodes, the same as very small areas. This PR restores this rendering by setting way_pixels to 0 if way_area is null, so points are treated the same as tiny polygons. Fixes #3876
and #3868 and related bugs.