Move parking back to amenity-points layers
Parking had previously been rendered in both amenity-points and amenity-low-priority.
Recently PR #3874 had consolidated parking rendering in low-priority only, but this leads to problems when there is a building or an address on the same feature
This commit moves parking back to the amenity-points layer, but near the end. The minzoom for text-low-priority is now changed back to z17