Change parking text intial zoom to z14
In PR #3612 it was planned to render parking from z14 only, however this was not changed for the text labels, only for the icon.
This commit fixes the parking rendering at z10 to z13, though in practice almost no parking lots are large enough to have been rendered at these levels.