Create darker river-color for river & canal areas and ditch, drain, stream, canal & river lines
Change river color to #8fcadd - Lch(78,21,227)
The color of river, canal, ditch and drain lines is changed to `@river-color` in `#8fcadd` to be more visible over forest and other dark landcover areas.
The color of river and canal areas is changed to match. This also improves mapper feedback about tagging of water areas
Added a new intermittent_river.png pattern in the new river-color