Render foot/cyclepaths without regard to surface
Rendering with a different dash pattern to indicate paved, unpaved, or
unknown was a good idea, but in practice the difference between them is
too subtle to be seen and understood.

The existing rendering also constrained us cartographically, making it
near-impossible to balance the patterns against a variety of backgrounds.