Render landuse=flowerbed
landuse=flowerbed is a gap in the rendering of urban vegetation. 13k occurrences according to Taginfo. I propose using the same rendering as landuse=orchard, as orchard and flowerbed are unlikely to be adjacent. I propose starting at zoom 11 because leisure=garden starts at zoom 10 and flowerbeds are smaller detail.