Use custom indexes to speed up ST_PointOnSurface queries
Using an index for ST_PointOnSurface queries speeds up the queries
substantially by avoiding work for when the geometries are in the
bounding box but the labeling point isn't.

The impact is substantial for large geometries like administrative
areas that have a lot of matches.

For these indexes to be reliably used, the plain way && !BBOX!
condition needs to be removed from the WHERE clause, making
the index required instead of optional.