Use the width of a runway for rendering, if set
Runways are fairly large features in the landscape, often highly visible
from above. The line rendering style of normal highways is not ideal
when zooming in on aerodromes, because the line representing the runway
diminishes into something much narrower than the actual area.

This commit introduces rendering based on the width tag of
`aeroway=runway`, and uses that to set the line-width. It defaults to
18 m if this tag is missing (this is on the narrow side, so it shouldn't
cause strange rendering where `width` is missing).

To help identify these large grey areas as runways, a centreline is
drawn.