De-emphasize parking=street_side (#4301)
For `amenity=parking` of the type `parking=street_side` (street-side
parking), the amenity icon is replaced with a smaller one that is shown
from z18 onwards.

With the introduction and formal acceptance of `parking=street_side`,
mappers now have a way of indicating that a mapped parking amenity is a
parking bay/area on the side of a street. This tag gives renderers a
chance to solve the problem of these parts of the street layout from
cluttering the map with large blue `P`'s that are better suited for
larger parking lots.

This commit introduces a smaller `P` icon that is pixel aligned to match
the crispness of the existing generic parking `P`.

Fixes #4262.