Improving hearing_aid symbol due to preliminary discussion
This version is based on number 28, with minor improvements and
svg optimisations (number 43). This symbol depicts the human ear
with an over-the-ear hearing aid being visually slightly separated
from the ear, as well as with sound waves propagating from behind
towards the hearing aid receiver.

Attention was given to a) uniqueness, b) visual simplicity, c)
rendering quality using kosmtik as well as png export with multiple
scaling factors (1x and 2x).