Flex version of the osm2pgsql configuration
This commit contains one file with an osm2pgsql configuration for the
flex output that can be used instead of the old configuration for the
pgsql output. It replaces the openstreetmap-carto.style and
openstreetmap-carto.lua files.

The configuration is nearly 100% compatible to the old one.

The database layout will be exactly the same with just very little
changes. The id columns (`osm_id`) and geometry columns (`way`) on all
tables will get the NOT NULL flag when using the flex output. These have
always been NOT NULL in practice anyway.

The content of the database will be the same with only minor irrelevant
differences.

Run like this:
osm2pgsql -O flex --style openstreetmap-carto-flex.lua -d gis ~/path/to/data.osm.pbf