Render `leisure=bleachers` ways.
Discussion held in
https://github.com/gravitystorm/openstreetmap-carto/issues/3140
suggested `leisure=bleachers` rendering could be similar in style to
pedestrian areas or other human-access features.

That is precisely what this patch does.

Colors were chosen to be those of `highway=pedestrian` because they are
pretty neutral; as `leisure=bleachers` can be used in a number of
contexts (such as sport pitches and outdoor theaters), I believe this
reduces potential visual confusion.

Note that it only renders ways; though found in the wild,
`leisure=bleachers` nodes are not supposed to be used.