Update list of ignored keys and key prefixes
I checked all keys that are in the planet file at least 1 million times
and added some uninteresting keys (usually ids from imports) to the
list.

Removed a few that are either not in the database any more or where
there are only a few objects left with that key.