prevent running out of memory with glavier rendering in PDF / SVG
I originally filed this as

https://github.com/mapnik/mapnik/issues/4476

seeing that rendering maps of certain areas in Greenland failed
with out-of-memory errors on print.get-map.org

Not an issue when rendering PNG tiles as there memory usage is
pretty much fixed to the actual output surface size; but with
PDF and SVG vector formats unclipped rendering of the Greenland
main ice sheet with a dashed line created lots of output outside
the actually visible area, eventually exhausting memory.

Adding clipping reduces the actual output to the visible bounding
box and so drastically reduces memory consumption.